requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==3.2.19
gunicorn==21.2.0
idna==3.4
packaging==23.1
paystack==1.5.0
Pillow==11.3.0
psycopg2==2.9.7
python-dotenv==0.21.1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
